Fix `docker` fact with recent version of docker
The `docker` nested fact is supposed to be conditionaly build if the
docker client version is 1.13 or greater, as changed in #166.

Unfortunately, the regexp is not working as expected, and future tweaks
of the regexp did not fix it.

Fix the regexp to match versions greater are equal to 1.13.*.  Add a few
unit tests to verify the outcome.

Fixes #896